move TaskRunner to a module within spartan. probably belongs in director later
