+++ modules/spartan/perception/handeyecalibration.py
@@ -3,8 +3,6 @@ import numpy as np
 import time
 import os
 import subprocess
-import sys
-from threading import Thread
 
 #director
 from director import lcmUtils
@@ -23,7 +21,6 @@ from director import drcargs
 from director import visualization as vis
 from director import ikplanner
 from director import robotposegui
-from director import asynctaskqueue
 from director.timercallback import TimerCallback
 
 
@@ -40,6 +37,7 @@ except:
 
 # spartan
 import spartan.utils as spartanUtils
+from spartan.taskrunner import TaskRunner
 
 """
 To run this set useKukaRLGDev to True in iiwaManipApp.py. This loads a
@@ -62,49 +60,6 @@ needed to run an AX=XB style hand-eye calibration.
 
 
 """
-
-
-class TaskRunner(object):
-
-  def __init__(self):
-    self.interval = 1/60.0
-    sys.setcheckinterval(1000)
-    #sys.setswitchinterval(self.interval)				# this is the only diff from Pat's snippet
-    													# sys.setswitchinterval is only Python 3
-    self.task_queue = asynctaskqueue.AsyncTaskQueue()
-    self.pending_tasks = []
-    self.threads = []
-    self.timer = TimerCallback(callback=self._on_timer)
-
-  def _on_timer(self):
-    time.sleep(self.interval)
-    if self.pending_tasks:
-      while True:
-        try:
-          self.task_queue.addTask(self.pending_tasks.pop(0))
-        except IndexError:
-          break
-
-      if self.task_queue.tasks and not self.task_queue.isRunning:
-        self.task_queue.start()
-
-    for t in self.threads:
-      if t.is_alive():
-        break
-    else:
-      self.threads = []
-      self.timer.stop()
-
-  def call_on_main(self, func, *args, **kwargs):
-    self.pending_tasks.append(lambda: func(*args, **kwargs))
-
-  def call_on_thread(self, func, *args, **kwargs):
-    t = Thread(target=lambda: func(*args, **kwargs))
-    self.threads.append(t)
-    t.start()
-    self.timer.targetFps = 1/self.interval
-    self.timer.start()
-
 
 class RobotService(object):
